# data-ref/src/main/resources/META-INF/resources/openapi.yaml
openapi: 3.0.3
info:
  title: API Baseline
  description: CRUD Operation on repositories data
  version: 0.7.0
servers:
  - url: http://data-ref.provoly.com
paths:
...
  /datasets/id/{id}:
    get:
      tags:
        - Dataset Controller
      operationId: com.provoly.ref.dataset.DatasetController_get
      summary: Retrieve a dataset based on its identifier.
      description: Also returns the metadata attached to the dataset, if any.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          $ref: '#/components/responses/datasetDetailsDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_read
            - search
    delete:
      tags:
        - Dataset Controller
      summary: Delete a dataset based on its identifier
      description: Deleting a dataset is not possible if dataset-version are associated with it
      operationId: com.provoly.ref.dataset.DatasetController_delete
      parameters:
        - name: id
          in: path
          description: dataset UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "400":
          description: Bad request
          content:
            application/json:
              examples:
                deleteDataset:
                  summary: Delete dataset with dataset-version
                  description: Occurs when you try to delete a dataset who owns dataset-version
                  value: |
                    {
                      "application": "data-ref",
                      "code": "BAD_REQUEST",
                      "message": "You're not allowed to delete dataset {{datasetId}} because it owns one or more dataset version"
                    }
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_write
...
  /datasets/id/{datasetId}/metadata/id/{metadataDefId}:
    put:
      tags:
        - Dataset Controller
      operationId: com.provoly.ref.dataset.DatasetController_setMetadata
      summary: Add metadata to a dataset
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: metadataDefId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MetadataValueWriteDto'
            examples:
              add_metadata:
                summary: Adding metadata to a dataset
                value: |
                  {
                    "value": "22"
                  }
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          $ref: '#/components/responses/wrongMetadataType'
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_write
    delete:
      tags:
        - Dataset Controller
      operationId: com.provoly.ref.dataset.DatasetController_deleteMetadata
      summary: Delete metadata of a dataset
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: metadataDefId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_write
  /datasets/class/{id}:
    get:
      tags:
        - Dataset Controller
      summary: Retrieve all datasets associated to a class
      operationId: com.provoly.ref.dataset.DatasetController_getAllForClass
      parameters:
        - name: id
          in: path
          description: Class UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          $ref: '#/components/responses/datasetDetailsDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - class_read
            - search
...
  /dataset-versions:
    post:
      tags:
        - Dataset Version Controller
      operationId: com.provoly.ref.dataset.DatasetVersionController_create
      summary: Save a new dataset version
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetVersionDto'
            examples:
              inactive:
                summary: Adding an inactive dataset version
                value: |
                  {
                    "id": "b1a08d9d-4c38-4d43-90f3-25015537ebfd" ,
                    "dataset": "423b5c01-1816-41d4-b358-200000000010",
                    "oClass": "991eb4b2-253c-4875-b708-8209fa5bfd51",
                    "lastModified": "2023-04-27T13:56:13.861018Z",
                    "version": 3,
                    "state": "INACTIVE",
                    "withFile": false
                  }
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          content:
            application/json:
              examples:
                idSaved:
                  summary: Create dataset version without unique id
                  description: Occurs when you try to create a dataset version of a closed dataset with an id already saved
                  value: |
                    {
                      "application": "data-ref",
                      "code": "BAD_REQUEST",
                      "message": "Dataset version {{datasetVersionId}} already exist."
                    }
                multipleVersionOnOpen:
                  summary: Create dataset version on an open dataset
                  description: Occurs when you try to create a second dataset version of an open dataset
                  value: |
                    {
                      "application": "data-ref",
                      "code": "BAD_REQUEST",
                      "message": "Dataset {{datasetId}} can't have more than one version."
                    }
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "409":
          description: Conflict
          content:
            application/json:
              examples:
                multipleOnClosedDataset:
                  summary: Create dataset version on a closed dataset
                  description: |
                    Occurs when you try to create a new dataset version on a closed dataset. This can also happen if you try to import data and an import is in progress on this dataset.
                  value: |
                    {
                      "application": "data-ref",
                      "code": "CONFLICT",
                      "message": "Can't override existing dataset {{datasetVersionId}} or have parallel imports"
                    }
      security:
        - oidc:
            - item_write
    put:
      tags:
        - Dataset Version Controller
      operationId: com.provoly.ref.dataset.DatasetVersionController_updateState
      summary: Update the state of a dataset version
      description: |
        A dataset version has multiple states to indicate the level of availability of the data imported.
        Each state can be modified to a specific set of other states. For example, if a dataset's state is `LOADING`,
        it can only obtain the `INDEXING` or `ERROR` states.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetVersionDto'
            examples:
              updateDatasetVersion:
                summary: Update dataset version
                value: |
                  {
                    "id": "b1a08d9d-4c38-4d43-90f3-25015537ebfd" ,
                    "dataset": "423b5c01-1816-41d4-b358-200000000010",
                    "oClass": "991eb4b2-253c-4875-b708-8209fa5bfd51",
                    "lastModified": "2023-04-27T13:56:13.861018Z",
                    "version": 1,
                    "state": "INACTIVE",
                    "withFile": false
                  }
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              examples:
                incompatibleStates:
                  summary: Wrong transition of state
                  description: Occurs when you try to update a dataset version's state to a state it can't transition to
                  value: |
                    {
                      "application": "data-ref",
                      "code": "BAD_REQUEST",
                      "message": "Dataset cannot transition from LOADING to ACTIVE"
                    }
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
          content:
            application/json:
              examples:
                differentDatasetId:
                  summary: Different dataset id
                  description: Occurs when you try to update a dataset version's state with a different dataset as the one in the database
                  value: |
                    {
                      "application": "data-ref",
                      "code": "FORBIDDEN",
                      "message": "Dataset versions are immutable"
                    }
        "409":
          description: Conflict
          content:
            application/json:
              examples:
                sameState:
                  summary: Update state with the actual state
                  description: Occurs when you try to update a dataset version's state with the same state it already is in
                  value: |
                    {
                      "application": "data-ref",
                      "code": "CONFLICT",
                      "message": "Dataset {{datasetId}} is already in state: {{datasetState}}"
                    }
      security:
        - oidc:
            - item_write
    get:
      tags:
        - Dataset Version Controller
      summary: Retrieve all dataset versions
      description: |
        A dataset version is an instantiation of a dataset. It allows us to have coherent and versioned class data. 
        The parameter hasWarnings in the response body is a boolean used to indicate if dataset-version import's generated some warnings.
      operationId: com.provoly.ref.dataset.DatasetVersionController_getAll
      responses:
        "200":
          $ref: '#/components/responses/listdatasetVersionDetailsDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_read
  /dataset-versions/id/{id}:
    get:
      tags:
        - Dataset Version Controller
      operationId: com.provoly.ref.dataset.DatasetVersionController_get
      summary: Retrieve a dataset version based on its identifier. It also returns its associated metadata.
      parameters:
        - name: id
          in: path
          description: dataset version UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          $ref: '#/components/responses/datasetVersionDetailsDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_read
            - search
    delete:
      tags:
        - Dataset Version Controller
      operationId: com.provoly.ref.dataset.DatasetVersionController_delete
      summary: Delete a dataset version based on its identifier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetVersionDto'
              examples:
                deletingDataset:
                  summary: Deleting dataset version
                  description: Return the dataset version currently deleting, this is an asynchrone operation.
                  value: |
                    {
                      "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "dataset": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "oClass": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "lastModified": "2022-03-10T16:15:50Z",
                      "version": 1,
                      "state": "DELETING",
                      "withFile": false
                    }
        "400":
          description: Bad Request
          content:
            application/json:
              examples:
                wrongState:
                  summary: Delete a version dataset in the wrong state
                  description: Occurs when you try to delete a dataset version in state INDEXING, LOADING, DELETING
                  value: |
                    {
                      "application": "data-ref",
                      "code": "BAD_REQUEST",
                      "message": "Unable to delete dataset version {{dataset_version_id}} because it's {{current_state}}"
                    }
        "401":
          description: Not Authorized
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_write
  /dataset-versions/id/{datasetVersionId}/metadata/id/{metadataDefId}:
    put:
      tags:
        - Dataset Version Controller
      operationId: com.provoly.ref.dataset.DatasetVersionController_setMetadata
      summary: Add metadata to a dataset version
      parameters:
        - name: datasetVersionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: metadataDefId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MetadataValueWriteDto'
            examples:
              default:
                value: |
                  {
                    "value": "22"
                  }
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          $ref: '#/components/responses/wrongMetadataType'
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_write
    delete:
      tags:
        - Dataset Version Controller
      operationId: com.provoly.ref.dataset.DatasetVersionController_deleteMetadata
      summary: Delete the metadata related to a dataset version
      parameters:
        - name: datasetVersionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: metadataDefId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request
          content:
            application/json:
              examples:
                metadataNotAssigned:
                  summary: Not assigned metadata
                  description: Occurs when you try to delete a metadata not assigned to the dataset version
                  value: |
                    {
                      "application": "data-ref",
                      "code": "BAD_REQUEST",
                      "message": "Metadata {{metadata_id}} is not assigned to DATASET_VERSION {{dataset_version_id}}"
                    }
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_write
  /dataset-versions/id/{id}/activate:
    post:
      tags:
        - Dataset Version Controller
      operationId: com.provoly.ref.dataset.DatasetVersionController_activate
      summary: Activate a dataset version based on its identifier
      description: |
        It also deactivates the previous one. 
        It is only possible to have only one version of an active dataset.
      parameters:
        - name: id
          in: path
          description: dataset version UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          content:
            application/json:
              examples:
                incompatibleStates:
                  summary: Incompatible transition states
                  description: Occurs when you try to activate a dataset version that is not ready to be activated
                  value: |
                    {
                      "application": "data-ref",
                      "code": "BAD_REQUEST",
                      "message": "Dataset cannot transition from LOADING to ACTIVE"
                    }
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
        "409":
          description: Conflict
          content:
            application/json:
              examples:
                alreadyActive:
                  summary: Already active
                  description: Occurs when you try to activate a dataset version that is already active
                  value: |
                    {
                      "application": "data-ref",
                      "code": "CONFLICT",
                      "message": "Dataset {{datasetId}} is already in state: ACTIVE"
                    }
      security:
        - oidc:
            - dataset_read
  /dataset-versions/id/{id}/deactivate:
    post:
      tags:
        - Dataset Version Controller
      operationId: com.provoly.ref.dataset.DatasetVersionController_deactivate
      summary: Deactivate a dataset version based on its identifier
      parameters:
        - name: id
          in: path
          description: dataset version UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          content:
            application/json:
              examples:
                incompatibleStates:
                  summary: Incompatible transition states
                  description: Occurs when you try to deactivate a dataset version that is not active
                  value: |
                    {
                      "application": "data-ref",
                      "code": "BAD_REQUEST",
                      "message": "Dataset cannot transition from ERROR to INACTIVE"
                    }
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
        "409":
          description: Conflict
          content:
            application/json:
              examples:
                alreadyInactive:
                  summary: Already inactive
                  description: Occurs when you try to deactivate a dataset version that is already inactive
                  value: |
                    {
                      "application": "data-ref",
                      "code": "CONFLICT",
                      "message": "Dataset {{datasetId}} is already in state: INACTIVE"
                    }

      security:
        - oidc:
            - dataset_read
  /dataset-versions/class/{classId}:
    get:
      tags:
        - Dataset Version Controller
      operationId: com.provoly.ref.dataset.DatasetVersionController_getAllForClass
      summary: Retrieve all dataset versions related to the class contained in its dataset
      parameters:
        - name: classId
          in: path
          description: class UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          $ref: '#/components/responses/listdatasetVersionDetailsDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - search
  /dataset-versions/id/{datasetVersionId}/previews:
    get:
      tags:
        - Dataset Version Controller
      operationId: com.provoly.ref.dataset.DatasetVersionController_getDatasetVersionPreviews
      summary: Returns the details of errors and warnings for a dataset version
      parameters:
        - name: datasetVersionId
          in: path
          description: datasetVersion UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: OK
          content:
            application/json:
              examples:
                datasetVersionPreviews:
                  summary: Dataset version errors and warnings
                  value: |
                    [
                      {
                        "level": "ERROR",
                        "messages": [
                          {
                            "level": "ERROR",
                            "extractMessageCode": "NO_ATTRIBUTES",
                            "type": null,
                            "name": null,
                            "recordId": null,
                            "datasetVersionId": "aba17aa9-fe0f-4610-8f36-60a244e875bb"
                          }
                        ],
                        "count": 1
                      },
                      {
                        "level": "WARNING",
                        "messages": [
                          {
                            "level": "WARNING",
                            "extractMessageCode": "UNRECOGNIZED",
                            "type": null,
                            "name": "Geo Point",
                            "recordId": null,
                            "datasetVersionId": "ff8dc620-57b4-4132-8326-efb01daac5b9"
                          },
                          {
                            "level": "WARNING",
                            "extractMessageCode": "UNRECOGNIZED",
                            "type": null,
                            "name": "Geo Shape",
                            "recordId": null,
                            "datasetVersionId": "ff8dc620-57b4-4132-8326-efb01daac5b9"
                          }
                        ],
                        "count": 2
                      }
                    ]
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_read
...
components:
  responses:
    notFoundException:
      description: Entity not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    wrongMetadataType:
      description: Not Allowed
      content:
        application/json:
          examples:
            wrongMetadataType:
              summary: Add metadata with unknown type
              description: Occurs when you try to add metadata with a type not supported
              value: |
                {
                  "application": "data-ref",
                  "code": "FORBIDDEN",
                  "message": "Metadata value should be of type : {{type}}"
                }
    datasetDetailsDto:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DatasetDetailsDto'
          examples:
            dataset:
              summary: Get a dataset with metadata
              value: |
                {
                  "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "oClass": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "type": "OPEN",
                  "name": "open dataset of my class",
                  "metadata": [
                    {
                      "value": 2,
                      "metadataDef": {
                        "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                        "name": "metadata_name",
                        "type": "INTEGER",
                        "description": "description of metadata"
                      }
                    }
                  ]
                }
    listdatasetVersionDetailsDto:
      description: OK
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/DatasetVersionDetailsDto'
          examples:
            getAllDatasetVersion:
              summary: Get list of dataset version disabled with warnings
              description: Example of a disabled dataset version that has warnings.
              value: |
                [
                  {
                    "id": "b1a08d9d-4c38-4d43-90f3-25015537ebfd" ,
                    "dataset": "423b5c01-1816-41d4-b358-200000000010",
                    "oClass": "991eb4b2-253c-4875-b708-8209fa5bfd51",
                    "lastModified": "2023-04-27T13:56:13.861018Z",
                    "version": 1,
                    "state": "INACTIVE",
                    "withFile": false,
                    "hasWarnings": true
                  }
                ]
            datasetVersionWithMetadata:
              summary: Get list of dataset version with metadata
              description: Example of a dataset version with metadata
              value: |
                [
                  {
                    "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "dataset": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "oClass": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "lastModified": "2022-03-10T16:15:50Z",
                    "version": 0,
                    "state": "LOADING",
                    "withFile": true,
                    "metadata": [
                      {
                        "value": 2,
                        "metadataDef": {
                          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                          "name": "metadata_name",
                          "type": "INTEGER",
                          "description": "description of metadata",
                          "slug": "string"
                        }
                      }
                    ]
                  }
                ]
    datasetVersionDetailsDto:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DatasetVersionDetailsDto'
          examples:
            getAllDatasetVersion:
              summary: Get dataset version with warnings
              description: Example of a dataset version that has warnings.
              value: |
                {
                  "id": "b1a08d9d-4c38-4d43-90f3-25015537ebfd" ,
                  "dataset": "423b5c01-1816-41d4-b358-200000000010",
                  "oClass": "991eb4b2-253c-4875-b708-8209fa5bfd51",
                  "lastModified": "2023-04-27T13:56:13.861018Z",
                  "version": 1,
                  "state": "ACTIVE",
                  "withFile": false,
                  "hasWarnings": true
                }
            datasetVersionWithMetadata:
              summary: Get dataset version with metadata
              description: Example of a dataset version with metadata
              value: |
                {
                  "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "dataset": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "oClass": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "lastModified": "2022-03-10T16:15:50Z",
                  "version": 2,
                  "state": "ACTIVE",
                  "withFile": true,
                  "metadata": [
                    {
                      "value": 2,
                      "metadataDef": {
                        "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                        "name": "metadata_name",
                        "type": "INTEGER",
                        "description": "description of metadata",
                        "slug": "string"
                      }
                    }
                  ]
                }
  schemas:
...
    DatasetDetailsDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "my dataset"
        oClass:
          $ref: '#/components/schemas/UUID'
        type:
          $ref: '#/components/schemas/DatasetType'
        metadata:
          type: array
          items:
            $ref: '#/components/schemas/MetadataValueReadDto'
        groups:
          type: array
          items:
            type: string
          example: [ "first_group", "second_group" ]
        description:
          type: string
          example: "my description"
        owner:
          type: boolean
          example: false
        slug:
          type: string
...
    DatasetVersionDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        dataset:
          $ref: '#/components/schemas/UUID'
        oClass:
          $ref: '#/components/schemas/UUID'
        lastModified:
          $ref: '#/components/schemas/Instant'
        version:
          format: int32
          type: integer
        state:
          $ref: '#/components/schemas/DatasetState'
        withFile:
          type: boolean
          example: false
...
    Instant:
      format: date-time
      type: string
      example: 2022-03-10T16:15:50Z
...
    MetadataValueWriteDto:
      type: object
      properties:
        value:
          type: string
        metadataDefId:
          $ref: '#/components/schemas/UUID'
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
  securitySchemes:
    oidc:
      type: openIdConnect
      description: Authentication
      openIdConnectUrl: https://auth/realms/provoly/.well-known/openid-configuration